backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "chess>=1.11.2",
    "fastapi[standard]",
    "game",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = ["ruff"]

[tool.uv.sources]
game = { workspace = true }
